+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Compatible with Chalk-2.0.2
+
 ## [1.0.0] - 2024-05-21
 
 ### Added